\name{ComparisonSurv}
\docType{package}
\alias{ComparisonSurv-package}
\title{Comparison of survival curves between two groups}
\description{Various statistical methods for survival analysis in comparing survival curves between two groups, including overall hypothesis tests described in Li et al. (2015), fixed-point tests in Klein et al. (2007), short-term tests, and long-term tests in Logan et al. (2008). Some commonly used descriptive statistics and plots are also included.
}
\author{Jingjing Lyu, Zheng Chen, Huimin Li, Jinbao Chen, Xinghui Huang.
maintainer: Zheng Chen \email{zheng-chen@hotmail.com}}
\references{
[1] Li H, Han D, Hou Y, et al. Statistical inference methods for two crossing survival curves: a comparison of methods. PLoS One, 2015, 10: e116774.

[2] Huang X, Lyu J, Hou Y, et al. A Nonparametric Statistical Method for Two Crossing Survival Curves. Communications in Statistics - Simulation and Computation, 2020. doi: 10.1080/03610918.2020.1753075.

[3] Klein JP, Logan B, Harhoff M, et al. Analyzing survival curves at a fixed point in time. Statistics in Medicine, 2007, 26(24):4505-4519.

[4] Logan BR, Klein JP, Zhang M. Comparing Treatments in the Presence of Crossing Survival Curves: An Application to Bone Marrow Transplantation. Biometrics, 2008, 64(3): 733-740.
}
